DEV-37860; refactor: move ai service types to separate file
